All notes and comments are resolved
Documentation for the file has been created.

Unit testing has been created with all possible test cases

NOTE: Typedef are not used in c++ se we can remove them

NOTE: For testing we can use arguments to perform testing procedure

IMPORTANT: Do not terminate the program using exit and abort, instead return the correct value for the mistake like -1 for failure and 0 for correctness to be able to monitor where the problem is

IMPORTANT: Do not use empty operators, maybe we can misunderstand that next line is in the loop (ex. while (getchar() != '\n'); )

IMPORTANT: #endif // !EQUATIONS_IO_H

IMPORTANT: Do not initialize variables in struct in c "COMPILATION ERROR"

IMPORTANT: Better not to use defines in main part of programs

QUESTION: in "assert(a != nullptr);" when I enter a as 0, I assert would be called and that's because NULL is actually 0

1) exit(-1) removed
2) cycle till get correct input
3) number_of_roots -> n_roots
4) SolveEquations -> SolveEquation
6) Move RootNumbers to another header
7) Use text file for inputs of tests and answers, to refuse more recompiling